Get nextsim running. Move classes, add includes.
