+++ core/src/modules/include/DevGrid.hpp
@@ -10,6 +10,7 @@
 
 #include "include/IStructure.hpp"
 
+#include "include/IDevGridIO.hpp"
 #include "include/ElementData.hpp"
 #include "include/PrognosticData.hpp"
 
@@ -36,13 +37,14 @@ public:
     }
 
     const static int nx;
+    const static std::string structureName;
 
     // Read/write override functions
     void init(const std::string& filePath) override;
 
     void dump(const std::string& filePath) const override;
 
-    std::string structureType() const override { return ourStructureName; };
+    std::string structureType() const override { return structureName; };
 
     // Cursor manipulation override functions
     int resetCursor() override;
@@ -51,47 +53,10 @@ public:
     const ElementData& cursorData() const override;
     void incrCursor() override;
 
-    /*!
-     * @brief A class that deals with all the netCDF related parts of DevGrid.
-     *
-     * @details DevGrid is a module, and currently merely mentioning it in the
-     * code requires the cpp file to be linked. Since this would pull in the
-     * NetCDF libraries, the DevGrid::IDevGridIO interface was created to
-     * separate out all the code that actually uses the NetCDF libraries. See
-     * DevGridIO for the implementing class.
-     */
-    class IDevGridIO {
-    public:
-        IDevGridIO(DevGrid& grid)
-            : grid(&grid)
-        {
-        }
-        virtual ~IDevGridIO() = default;
-        /*!
-         * @brief Reads data from the file location into the vector of data elements.
-         *
-         * @param dg The vector of ElementData instances to be filled.
-         * @param filePath The location of the NetCDF restart file to be read.
-         */
-        virtual void init(std::vector<ElementData>& dg, const std::string& filePath) const = 0;
-        /*!
-         * @brief Writes data from the vector of data elements into the file location.
-         *
-         * @param dg The vector of ElementData instances containing the data.
-         * @param filePath The location of the NetCDF restart file to be written.
-         */
-        virtual void dump(
-            const std::vector<ElementData>& dg, const std::string& fielPath) const = 0;
-
-    protected:
-        DevGrid* grid;
-    };
-
     //! Sets the pointer to the class that will perform the IO. Should be an instance of DevGridIO
     void setIO(IDevGridIO* p) { pio = p; }
 
 private:
-    const static std::string ourStructureName;
     const static std::string xDimName;
     const static std::string yDimName;
